fix(fleet-carriers): keep the final report parser off the package root
The archive module is re-exported both as a wildcard and as the jobs.archive
namespace, so folding report-extract.ts into it published extractFinalReport
from the package root. Restore the parser to its own non-barrel module; the
namespace then carries the same members it did before the consolidation.
